refactor: Betting 클래스 game 패키지로 이동
